Add cast_tests.cpp and move fixed_point tests
